backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
xgboost==2.0.3
scikit-learn==1.3.2
numpy==1.26.2
python-multipart==0.0.6
python-dotenv==1.0.1
twilio==8.10.0
email-validator==2.3.0
SQLAlchemy==2.0.36
PyMySQL==1.1.1
